// bajaj-events-portal/src/app/events/models/event.ts
export class Event {
        eventId: number;
        eventCode: string;
        eventName: string;
        description: string;
        startDate: Date;
        endDate: Date;
        fees: number;
        seatsFilled: number;
        logo: string;
        venue: string;
        organizedBy: string;
}